+++ internal/xerrors/transport.go
@@ -15,30 +15,25 @@ import (
 )
 
 type transportError struct {
-	code    grpcCodes.Code
-	message string
-	err     error
-	details []interface{}
+	status  *grpcStatus.Status
 	address string
+}
+
+func (e *transportError) GRPCStatus() *grpcStatus.Status {
+	return e.status
 }
 
 func (e *transportError) isYdbError() {}
 
 func (e *transportError) Code() int32 {
-	return int32(e.code)
+	return int32(e.status.Code())
 }
 
 func (e *transportError) Name() string {
-	return "transport/" + e.code.String()
+	return "transport/" + e.status.Code().String()
 }
 
 type teOpt func(te *transportError)
-
-func WithCode(code grpcCodes.Code) teOpt {
-	return func(te *transportError) {
-		te.code = code
-	}
-}
 
 func WithAddress(address string) teOpt {
 	return func(te *transportError) {
@@ -46,34 +41,23 @@ func WithAddress(address string) teOpt {
 	}
 }
 
-// Transport returns a new transport error with given options
-func Transport(opts ...teOpt) error {
-	te := &transportError{}
-	for _, f := range opts {
-		if f != nil {
-			f(te)
-		}
-	}
-	return WithStackTrace(fmt.Errorf("%w", te), WithSkipDepth(1))
-}
-
 func (e *transportError) Error() string {
 	var b bytes.Buffer
 	b.WriteString("transport error: ")
-	b.WriteString(e.code.String())
-	if e.message != "" {
+	b.WriteString(e.status.Code().String())
+	if message := e.status.Message(); message != "" {
 		b.WriteString(", message: ")
-		b.WriteString(e.message)
+		b.WriteString(message)
 	}
 	if len(e.address) > 0 {
 		b.WriteString(", address: ")
 		b.WriteString(e.address)
 	}
-	if len(e.details) > 0 {
+	if details := e.status.Details(); len(details) > 0 {
 		b.WriteString(", details: ")
-		if len(e.details) > 0 {
+		if len(details) > 0 {
 			b.WriteString(", details:")
-			for _, detail := range e.details {
+			for _, detail := range details {
 				b.WriteString(fmt.Sprintf("\n- %v", detail))
 			}
 		}
@@ -82,7 +66,7 @@ func (e *transportError) Error() string {
 }
 
 func (e *transportError) Unwrap() error {
-	return e.err
+	return e.status.Err()
 }
 
 func dumpIssues(buf *bytes.Buffer, ms []*Ydb_Issue.IssueMessage) {
@@ -106,7 +90,7 @@ func dumpIssues(buf *bytes.Buffer, ms []*Ydb_Issue.IssueMessage) {
 }
 
 func (e *transportError) Type() Type {
-	switch e.code {
+	switch e.status.Code() {
 	case
 		grpcCodes.Aborted,
 		grpcCodes.ResourceExhausted:
@@ -122,7 +106,7 @@ func (e *transportError) Type() Type {
 }
 
 func (e *transportError) BackoffType() backoff.Type {
-	switch e.code {
+	switch e.status.Code() {
 	case
 		grpcCodes.Internal,
 		grpcCodes.Canceled,
@@ -136,7 +120,7 @@ func (e *transportError) BackoffType() backoff.Type {
 }
 
 func (e *transportError) MustDeleteSession() bool {
-	switch e.code {
+	switch e.status.Code() {
 	case
 		grpcCodes.ResourceExhausted,
 		grpcCodes.OutOfRange:
@@ -158,7 +142,7 @@ func IsTransportError(err error, codes ...grpcCodes.Code) bool {
 			return true
 		}
 		for _, code := range codes {
-			if t.code == code {
+			if t.status.Code() == code {
 				return true
 			}
 		}
@@ -168,7 +152,8 @@ func IsTransportError(err error, codes ...grpcCodes.Code) bool {
 	}
 }
 
-func FromGRPCError(err error, opts ...teOpt) error {
+// Transport returns a new transport error with given options
+func Transport(err error, opts ...teOpt) error {
 	if err == nil {
 		return nil
 	}
@@ -178,10 +163,7 @@ func FromGRPCError(err error, opts ...teOpt) error {
 	}
 	if s, ok := grpcStatus.FromError(err); ok {
 		te := &transportError{
-			code:    s.Code(),
-			message: s.Message(),
-			err:     s.Err(),
-			details: s.Details(),
+			status: s,
 		}
 		for _, o := range opts {
 			if o != nil {