+++ retry/errors_data_test.go
@@ -7,7 +7,7 @@ import (
 	"github.com/ydb-platform/ydb-go-genproto/protos/Ydb"
 	"google.golang.org/grpc"
 	grpcCodes "google.golang.org/grpc/codes"
-	"google.golang.org/grpc/status"
+	grpcStatus "google.golang.org/grpc/status"
 
 	"github.com/ydb-platform/ydb-go-sdk/v3/internal/backoff"
 	"github.com/ydb-platform/ydb-go-sdk/v3/internal/xerrors"
@@ -64,7 +64,7 @@ var errsToCheck = []struct {
 		},
 	},
 	{
-		err: xerrors.FromGRPCError(
+		err: xerrors.Transport(
 			grpc.ErrClientConnClosing, //nolint:staticcheck
 		),
 		backoff:       backoff.TypeFast,
@@ -75,18 +75,7 @@ var errsToCheck = []struct {
 		},
 	},
 	{
-		err:           xerrors.Transport(),
-		backoff:       backoff.TypeNoBackoff,
-		deleteSession: true,
-		canRetry: map[idempotency]bool{
-			idempotent:    false,
-			nonIdempotent: false,
-		},
-	},
-	{
-		err: xerrors.Transport(
-			xerrors.WithCode(grpcCodes.Canceled),
-		),
+		err:           xerrors.Transport(grpcStatus.Error(grpcCodes.Canceled, "")),
 		backoff:       backoff.TypeFast,
 		deleteSession: true,
 		canRetry: map[idempotency]bool{
@@ -95,9 +84,7 @@ var errsToCheck = []struct {
 		},
 	},
 	{
-		err: xerrors.Transport(
-			xerrors.WithCode(grpcCodes.Unknown),
-		),
+		err:           xerrors.Transport(grpcStatus.Error(grpcCodes.Unknown, "")),
 		backoff:       backoff.TypeNoBackoff,
 		deleteSession: true,
 		canRetry: map[idempotency]bool{
@@ -106,9 +93,7 @@ var errsToCheck = []struct {
 		},
 	},
 	{
-		err: xerrors.Transport(
-			xerrors.WithCode(grpcCodes.InvalidArgument),
-		),
+		err:           xerrors.Transport(grpcStatus.Error(grpcCodes.InvalidArgument, "")),
 		backoff:       backoff.TypeNoBackoff,
 		deleteSession: true,
 		canRetry: map[idempotency]bool{
@@ -117,9 +102,7 @@ var errsToCheck = []struct {
 		},
 	},
 	{
-		err: xerrors.Transport(
-			xerrors.WithCode(grpcCodes.DeadlineExceeded),
-		),
+		err:           xerrors.Transport(grpcStatus.Error(grpcCodes.DeadlineExceeded, "")),
 		backoff:       backoff.TypeNoBackoff,
 		deleteSession: true,
 		canRetry: map[idempotency]bool{
@@ -128,9 +111,7 @@ var errsToCheck = []struct {
 		},
 	},
 	{
-		err: xerrors.Transport(
-			xerrors.WithCode(grpcCodes.NotFound),
-		),
+		err:           xerrors.Transport(grpcStatus.Error(grpcCodes.NotFound, "")),
 		backoff:       backoff.TypeNoBackoff,
 		deleteSession: true,
 		canRetry: map[idempotency]bool{
@@ -139,9 +120,7 @@ var errsToCheck = []struct {
 		},
 	},
 	{
-		err: xerrors.Transport(
-			xerrors.WithCode(grpcCodes.AlreadyExists),
-		),
+		err:           xerrors.Transport(grpcStatus.Error(grpcCodes.AlreadyExists, "")),
 		backoff:       backoff.TypeNoBackoff,
 		deleteSession: true,
 		canRetry: map[idempotency]bool{
@@ -150,9 +129,7 @@ var errsToCheck = []struct {
 		},
 	},
 	{
-		err: xerrors.Transport(
-			xerrors.WithCode(grpcCodes.PermissionDenied),
-		),
+		err:           xerrors.Transport(grpcStatus.Error(grpcCodes.PermissionDenied, "")),
 		backoff:       backoff.TypeNoBackoff,
 		deleteSession: true,
 		canRetry: map[idempotency]bool{
@@ -161,9 +138,7 @@ var errsToCheck = []struct {
 		},
 	},
 	{
-		err: xerrors.Transport(
-			xerrors.WithCode(grpcCodes.ResourceExhausted),
-		),
+		err:           xerrors.Transport(grpcStatus.Error(grpcCodes.ResourceExhausted, "")),
 		backoff:       backoff.TypeSlow,
 		deleteSession: false,
 		canRetry: map[idempotency]bool{
@@ -172,9 +147,7 @@ var errsToCheck = []struct {
 		},
 	},
 	{
-		err: xerrors.Transport(
-			xerrors.WithCode(grpcCodes.FailedPrecondition),
-		),
+		err:           xerrors.Transport(grpcStatus.Error(grpcCodes.FailedPrecondition, "")),
 		backoff:       backoff.TypeNoBackoff,
 		deleteSession: true,
 		canRetry: map[idempotency]bool{
@@ -183,9 +156,7 @@ var errsToCheck = []struct {
 		},
 	},
 	{
-		err: xerrors.Transport(
-			xerrors.WithCode(grpcCodes.Aborted),
-		),
+		err:           xerrors.Transport(grpcStatus.Error(grpcCodes.Aborted, "")),
 		backoff:       backoff.TypeNoBackoff,
 		deleteSession: true,
 		canRetry: map[idempotency]bool{
@@ -194,9 +165,7 @@ var errsToCheck = []struct {
 		},
 	},
 	{
-		err: xerrors.Transport(
-			xerrors.WithCode(grpcCodes.OutOfRange),
-		),
+		err:           xerrors.Transport(grpcStatus.Error(grpcCodes.OutOfRange, "")),
 		backoff:       backoff.TypeNoBackoff,
 		deleteSession: false,
 		canRetry: map[idempotency]bool{
@@ -205,9 +174,7 @@ var errsToCheck = []struct {
 		},
 	},
 	{
-		err: xerrors.Transport(
-			xerrors.WithCode(grpcCodes.Unimplemented),
-		),
+		err:           xerrors.Transport(grpcStatus.Error(grpcCodes.Unimplemented, "")),
 		backoff:       backoff.TypeNoBackoff,
 		deleteSession: true,
 		canRetry: map[idempotency]bool{
@@ -216,9 +183,7 @@ var errsToCheck = []struct {
 		},
 	},
 	{
-		err: xerrors.Transport(
-			xerrors.WithCode(grpcCodes.Internal),
-		),
+		err:           xerrors.Transport(grpcStatus.Error(grpcCodes.Internal, "")),
 		backoff:       backoff.TypeFast,
 		deleteSession: true,
 		canRetry: map[idempotency]bool{
@@ -227,9 +192,7 @@ var errsToCheck = []struct {
 		},
 	},
 	{
-		err: xerrors.Transport(
-			xerrors.WithCode(grpcCodes.Unavailable),
-		),
+		err:           xerrors.Transport(grpcStatus.Error(grpcCodes.Unavailable, "")),
 		backoff:       backoff.TypeFast,
 		deleteSession: true,
 		canRetry: map[idempotency]bool{
@@ -239,9 +202,7 @@ var errsToCheck = []struct {
 	},
 	{
 		err: xerrors.Retryable(
-			xerrors.Transport(
-				xerrors.WithCode(grpcCodes.Unavailable),
-			),
+			xerrors.Transport(grpcStatus.Error(grpcCodes.Unavailable, "")),
 			xerrors.WithBackoff(backoff.TypeFast),
 			xerrors.WithDeleteSession(),
 		),
@@ -254,7 +215,7 @@ var errsToCheck = []struct {
 	},
 	{
 		err: xerrors.Retryable(
-			status.Error(grpcCodes.Unavailable, ""),
+			grpcStatus.Error(grpcCodes.Unavailable, ""),
 			xerrors.WithBackoff(backoff.TypeFast),
 			xerrors.WithDeleteSession(),
 		),
@@ -266,9 +227,7 @@ var errsToCheck = []struct {
 		},
 	},
 	{
-		err: xerrors.Transport(
-			xerrors.WithCode(grpcCodes.DataLoss),
-		),
+		err:           xerrors.Transport(grpcStatus.Error(grpcCodes.DataLoss, "")),
 		backoff:       backoff.TypeNoBackoff,
 		deleteSession: true,
 		canRetry: map[idempotency]bool{
@@ -277,9 +236,7 @@ var errsToCheck = []struct {
 		},
 	},
 	{
-		err: xerrors.Transport(
-			xerrors.WithCode(grpcCodes.Unauthenticated),
-		),
+		err:           xerrors.Transport(grpcStatus.Error(grpcCodes.Unauthenticated, "")),
 		backoff:       backoff.TypeNoBackoff,
 		deleteSession: true,
 		canRetry: map[idempotency]bool{