+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 * Added checks for nil option to all opts range loops
 * Moved content of package `internal/ctxlabels` into `internal/xcontext`
+* Implemented `GRPCStatus` method in `internal/xerrors/transportError`
 
 ## v3.42.5
 * Fixed closing of `database/sql` connection (aka `YDB` session)